+++ test/args.jl
@@ -87,5 +87,8 @@
         # flags are always stored, even if not set
         "use-gurobi" => false,
         "opt-switch-formulation" => "lindistflow",
+        "opt-switch-algorithm" => "iterative",
+        "opt-switch-solver" => "misocp_solver",
+        "opt-disp-algorithm" => "opf",
     )
 end